backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2

# Agno and AI dependencies
agno
openai

# Database
sqlalchemy

# Optional tools (uncomment if needed)
# duckduckgo-search

# Environment and configuration
python-dotenv
